# web/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Go Web bj38web",
        "title": "bj38web",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "author：@ouzhsh",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "0.0.1"
    },
    "host": "127.0.0.1:8080",
    "basePath": "/api/v1.0",
    "paths": {
...
        "/api/v1.0/imagecode/:uuid": {
            "get": {
                "description": "获取图片验证码",
                "consumes": [
                    "text/plain"
                ],
                "produces": [
                    "image/png"
                ],
                "tags": [
                    "用户业务接口"
                ],
                "summary": "图片验证码",
                "parameters": [
                    {
                        "type": "string",
                        "description": "路径变量参数",
                        "name": "uuid",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "验证码",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
        "/api/v1.0/session": {
            "get": {
                "description": "获取session信息服务",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "用户业务接口"
                ],
                "summary": "用户会话",
                "parameters": [
                    {
                        "type": "string",
                        "description": "查询参数",
                        "name": "object",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "用户名",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "delete": {
                "description": "退出登录状态",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "用户业务接口"
                ],
                "summary": "退出登录状态",
                "parameters": [
                    {
                        "type": "string",
                        "description": "查询参数",
                        "name": "object",
                        "in": "formData"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "消息",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/v1.0/sessions": {
            "post": {
                "description": "用户登录",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "用户业务接口"
                ],
                "summary": "用户登录",
                "parameters": [
                    {
                        "description": "手机号",
                        "name": "mobile",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "description": "密码",
                        "name": "password",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "消息",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/v1.0/smscode/:mobile": {
            "get": {
                "description": "获取短信验证码",
                "consumes": [
                    "text/plain"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "用户业务接口"
                ],
                "summary": "短信验证码",
                "parameters": [
                    {
                        "type": "string",
                        "description": "路径变量参数",
                        "name": "phone",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "消息",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/v1.0/user": {
            "get": {
                "description": "获取用户信息",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "用户业务接口"
                ],
                "summary": "获取用户信息",
                "parameters": [
                    {
                        "type": "string",
                        "description": "用户名",
                        "name": "userName",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "用户信息",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/v1.0/user/auth": {
            "post": {
                "description": "用户实名认证",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "用户实名认证"
                ],
                "summary": "用户实名认证",
                "parameters": [
                    {
                        "description": "真实姓名",
                        "name": "real_name",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "description": "身份证",
                        "name": "id_card",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "信息",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/v1.0/user/avatar": {
            "post": {
                "description": "上传用户头像",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "上传用户头像"
                ],
                "summary": "上传用户头像",
                "parameters": [
                    {
                        "type": "file",
                        "description": "用户名",
                        "name": "avatar",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "信息",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/v1.0/user/houses": {
            "get": {
                "description": "获取用户发布的房源信息",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "获取用户发布的房源信息"
                ],
                "summary": "获取用户发布的房源信息",
                "parameters": [
                    {
                        "description": "用户名",
                        "name": "userName",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "信息",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/v1.0/user/name": {
            "put": {
                "description": "修改用户名提交",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "修改用户名提交"
                ],
                "summary": "修改用户名提交",
                "parameters": [
                    {
                        "description": "用户名",
                        "name": "name",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "用户信息",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
        "/api/v1.0/users": {
            "post": {
                "description": "提交用户注册请求",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "用户业务接口"
                ],
                "summary": "提交用户注册请求",
                "parameters": [
                    {
                        "description": "手机号",
                        "name": "mobile",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "description": "密码",
                        "name": "password",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "description": "验证码",
                        "name": "sms_code",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "消息",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        }
    },
...
}